# Fix series preview table height on the course page