--- a/app/assets/javascripts/course.ts
+++ b/app/assets/javascripts/course.ts
@@ -94,7 +94,7 @@ function initCourseMembers(): void {
 }
 
 const TABLE_WRAPPER_SELECTOR = ".series-activities-table-wrapper";
-const SKELETON_TABLE_SELECTOR = ".activity-table-skeletons";
+const SKELETON_TABLE_SELECTOR = ".activity-table-skeleton";
 
 class Series {
     private readonly id: number;