--- a/app/assets/stylesheets/models/series.css.scss
+++ b/app/assets/stylesheets/models/series.css.scss
@@ -180,20 +180,27 @@
   white-space: normal;
 }
 
-.skeleton-field {
-  background-image: linear-gradient($skeleton-color 100%, transparent 0);
-
-  &.skeleton-progress {
-    height: 8px;
-    width: 100px;
+.skeleton-table {
+  .sk-p {
+    background-image: linear-gradient($skeleton-color 100%, transparent 0);
+    width: 116px;
+    background-size: 100px 8px;
+    background-repeat: no-repeat;
+    background-position-y: center;
+    background-position-x: 8px;
   }
 
-  &.skeleton-status {
-    height: 17px;
-    width: 110px;
+  .sk-s {
+    background-image: linear-gradient($skeleton-color 100%, transparent 0);
+    width: 126px;
+    background-size: 110px 17px;
+    background-repeat: no-repeat;
+    background-position-y: center;
+    background-position-x: 8px;
   }
 
-  &.skeleton-icon {
+  .sk-i {
+    background-image: linear-gradient($skeleton-color 100%, transparent 0);
     display: block;
     background-size: 17px 17px;
     background-repeat: no-repeat;