--- a/app/models/submission.rb
+++ b/app/models/submission.rb
@@ -60,6 +60,8 @@ class Submission < ApplicationRecord
   scope :of_judge, ->(judge) { where(exercise_id: Exercise.where(judge_id: judge.id)) }
   scope :from_students, ->(course) { where(user: course.enrolled_members) }
 
+  # This scope used to use `where.not` with less options, but that was a lot slower because SQL does not use indexes in that case
+  # As this scope is used for the progress bars on the course page, the speedup was relevant
   scope :judged, -> { where(status: ['unknown', 'correct', 'wrong', 'time limit exceeded', 'runtime error', 'compilation error', 'memory limit exceeded', 'internal error', 'output limit exceeded']) }
   scope :by_exercise_name, ->(name) { where(exercise: Exercise.by_name(name)) }
   scope :by_status, ->(status) { where(status: status.in?(statuses) ? status : -1) }