--- a/db/migrate/20221117132148_add_exercise_status_course_index_to_submissions.rb
+++ b/db/migrate/20221117132148_add_exercise_status_course_index_to_submissions.rb
@@ -1,0 +1,5 @@
+class AddExerciseStatusCourseIndexToSubmissions < ActiveRecord::Migration[7.0]
+  def change
+    add_index :submissions, [:exercise_id, :status, :course_id], name: 'ex_st_co_idx'
+  end
+end