--- a/db/schema.rb
+++ b/db/schema.rb
@@ -10,13 +10,13 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
+ActiveRecord::Schema[7.0].define(version: 2022_11_17_132148) do
   create_table "active_storage_attachments", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.string "name", null: false
     t.string "record_type", null: false
     t.bigint "record_id", null: false
     t.bigint "blob_id", null: false
-    t.datetime "created_at", null: false
+    t.datetime "created_at", precision: nil, null: false
     t.index ["blob_id"], name: "index_active_storage_attachments_on_blob_id"
     t.index ["record_type", "record_id", "name", "blob_id"], name: "index_active_storage_attachments_uniqueness", unique: true
   end
@@ -28,12 +28,12 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.text "metadata", size: :medium
     t.bigint "byte_size", null: false
     t.string "checksum"
-    t.datetime "created_at", null: false
+    t.datetime "created_at", precision: nil, null: false
     t.string "service_name", null: false
     t.index ["key"], name: "index_active_storage_blobs_on_key", unique: true
   end
 
-  create_table "active_storage_variant_records", charset: "utf8mb4", force: :cascade do |t|
+  create_table "active_storage_variant_records", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
@@ -42,8 +42,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
   create_table "activities", id: :integer, charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.string "name_nl"
     t.string "name_en"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "path"
     t.string "description_format"
     t.integer "repository_id"
@@ -90,7 +90,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.boolean "accepted_before_deadline", default: false, null: false
     t.boolean "solved", default: false, null: false
     t.boolean "started", default: false, null: false
-    t.datetime "solved_at"
+    t.datetime "solved_at", precision: nil
     t.integer "activity_id", null: false
     t.integer "series_id"
     t.integer "user_id", null: false
@@ -104,7 +104,6 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.index ["activity_id"], name: "index_activity_statuses_on_activity_id"
     t.index ["series_id"], name: "fk_rails_1bc42c2178"
     t.index ["user_id", "series_id_non_nil", "activity_id"], name: "index_on_user_id_series_id_non_nil_activity_id", unique: true
-    t.index ["user_id"], name: "fk_rails_8a05a160e8"
   end
 
   create_table "annotations", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
@@ -128,7 +127,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.index ["user_id"], name: "index_annotations_on_user_id"
   end
 
-  create_table "announcement_views", charset: "utf8mb4", force: :cascade do |t|
+  create_table "announcement_views", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.bigint "announcement_id", null: false
     t.integer "user_id", null: false
     t.datetime "created_at", null: false
@@ -138,7 +137,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.index ["user_id"], name: "index_announcement_views_on_user_id"
   end
 
-  create_table "announcements", charset: "utf8mb4", force: :cascade do |t|
+  create_table "announcements", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.text "text_nl", null: false
     t.text "text_en", null: false
     t.datetime "start_delivering_at"
@@ -155,8 +154,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.bigint "user_id"
     t.string "token_digest"
     t.string "description"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["token_digest"], name: "index_api_tokens_on_token_digest"
     t.index ["user_id", "description"], name: "index_api_tokens_on_user_id_and_description", unique: true
     t.index ["user_id"], name: "index_api_tokens_on_user_id"
@@ -165,8 +164,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
   create_table "course_labels", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.integer "course_id", null: false
     t.string "name", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["course_id", "name"], name: "index_course_labels_on_course_id_and_name", unique: true
   end
 
@@ -181,8 +180,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.integer "course_id", null: false
     t.integer "user_id", null: false
     t.integer "status", default: 2, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.boolean "favorite", default: false
     t.index ["course_id"], name: "index_course_memberships_on_course_id"
     t.index ["user_id", "course_id"], name: "index_course_memberships_on_user_id_and_course_id", unique: true
@@ -200,11 +199,11 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "name"
     t.string "year"
     t.string "secret"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.text "description", size: :medium
-    t.integer "visibility"
-    t.integer "registration"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.text "description", size: :long
+    t.integer "visibility", default: 0
+    t.integer "registration", default: 0
     t.string "teacher"
     t.bigint "institution_id"
     t.string "search", limit: 4096
@@ -218,15 +217,15 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
   create_table "delayed_jobs", id: :integer, charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.integer "priority", default: 0, null: false
     t.integer "attempts", default: 0, null: false
-    t.text "handler", size: :medium, null: false
-    t.text "last_error", size: :medium
-    t.datetime "run_at"
-    t.datetime "locked_at"
-    t.datetime "failed_at"
+    t.text "handler", size: :long, null: false
+    t.text "last_error", size: :long
+    t.datetime "run_at", precision: nil
+    t.datetime "locked_at", precision: nil
+    t.datetime "failed_at", precision: nil
     t.string "locked_by"
     t.string "queue"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at", precision: nil
+    t.datetime "updated_at", precision: nil
     t.index ["priority", "run_at"], name: "delayed_jobs_priority"
   end
 
@@ -254,7 +253,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
   create_table "evaluations", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.integer "series_id"
     t.boolean "released", default: false, null: false
-    t.datetime "deadline", null: false
+    t.datetime "deadline", precision: nil, null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["series_id"], name: "index_evaluations_on_series_id"
@@ -264,8 +263,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.integer "event_type", null: false
     t.integer "user_id"
     t.string "message", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["event_type"], name: "index_events_on_event_type"
     t.index ["user_id"], name: "fk_rails_0cb5590091"
   end
@@ -309,8 +308,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "name"
     t.string "short_name"
     t.string "logo"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.boolean "generated_name", default: true, null: false
     t.integer "category", default: 0, null: false
   end
@@ -319,8 +318,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "name"
     t.string "image"
     t.string "path"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "renderer", null: false
     t.string "remote"
     t.integer "clone_status", default: 1, null: false
@@ -349,8 +348,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "name", null: false
     t.string "editor_name", null: false
     t.string "extension", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "icon"
     t.string "renderer_name", null: false
     t.index ["name"], name: "index_programming_languages_on_name", unique: true
@@ -380,8 +379,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "remote"
     t.string "path"
     t.integer "judge_id"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "clone_status", default: 1, null: false
     t.index ["judge_id"], name: "index_repositories_on_judge_id"
     t.index ["name"], name: "index_repositories_on_name", unique: true
@@ -395,7 +394,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.index ["user_id"], name: "fk_rails_6b59ad362c"
   end
 
-  create_table "rights_requests", charset: "utf8mb4", force: :cascade do |t|
+  create_table "rights_requests", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.integer "user_id", null: false
     t.string "institution_name"
     t.text "context", null: false
@@ -404,7 +403,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.index ["user_id"], name: "index_rights_requests_on_user_id"
   end
 
-  create_table "saved_annotations", charset: "utf8mb4", force: :cascade do |t|
+  create_table "saved_annotations", charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.string "title", null: false
     t.text "annotation_text", size: :medium
     t.integer "user_id", null: false
@@ -446,12 +445,12 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
   create_table "series", id: :integer, charset: "utf8mb4", collation: "utf8mb4_unicode_ci", force: :cascade do |t|
     t.integer "course_id"
     t.string "name"
-    t.text "description", size: :medium
+    t.text "description", size: :long
     t.integer "visibility"
     t.integer "order", default: 0, null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.datetime "deadline"
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
+    t.datetime "deadline", precision: nil
     t.string "access_token"
     t.string "indianio_token"
     t.boolean "progress_enabled", default: true, null: false
@@ -470,8 +469,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.integer "series_id"
     t.integer "activity_id"
     t.integer "order", default: 999
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.index ["activity_id"], name: "index_series_memberships_on_activity_id"
     t.index ["series_id", "activity_id"], name: "index_series_memberships_on_series_id_and_activity_id", unique: true
     t.index ["series_id"], name: "index_series_memberships_on_series_id"
@@ -481,8 +480,8 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.integer "exercise_id"
     t.integer "user_id"
     t.string "summary"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.integer "status"
     t.boolean "accepted", default: false
     t.integer "course_id"
@@ -490,6 +489,7 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.integer "number"
     t.index ["accepted"], name: "index_submissions_on_accepted"
     t.index ["course_id"], name: "index_submissions_on_course_id"
+    t.index ["exercise_id", "status", "course_id"], name: "ex_st_co_idx"
     t.index ["exercise_id", "user_id", "accepted", "created_at"], name: "ex_us_ac_cr_index"
     t.index ["exercise_id", "user_id", "status", "created_at"], name: "ex_us_st_cr_index"
     t.index ["exercise_id"], name: "index_submissions_on_exercise_id"
@@ -504,15 +504,15 @@ ActiveRecord::Schema[7.0].define(version: 2022_09_14_075029) do
     t.string "last_name"
     t.string "email"
     t.integer "permission", default: 0
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.datetime "created_at", precision: nil, null: false
+    t.datetime "updated_at", precision: nil, null: false
     t.string "lang", default: "nl"
     t.string "token"
     t.string "time_zone", default: "Brussels"
     t.bigint "institution_id"
     t.string "search", limit: 4096
-    t.datetime "seen_at"
-    t.datetime "sign_in_at"
+    t.datetime "seen_at", precision: nil
+    t.datetime "sign_in_at", precision: nil
     t.index ["email", "institution_id"], name: "index_users_on_email_and_institution_id", unique: true
     t.index ["institution_id"], name: "index_users_on_institution_id"
     t.index ["token"], name: "index_users_on_token"