This pull request is a third attempt to fix #3962. See #4167 and  #4169 for previous attempts.

This pull request contains two parts:
- A replacement of the old table preview. The new table preview has the correct height in all browser and also contains the activity name as requested by @niknetniko .
- A speed up of the progress bar load, to offset reduced speed by the new table.

## Analysis
To analyze the impact of these changes, I looked at two large courses. For fair comparison I disabled memcached and the sql query cache. 

### http://dodona.ugent.be/nl/courses/800/
102 series | 640 exercises | 488 users | 491.449 submissions

| Pr    | Total request time | Time in views | Time in Activerecord | Filesize |
|-------|--------------|---------------|----------------------|----------|
| Old   | 755ms        | 620.0ms       | 125.4ms              | 833KB    |
| #4167 | 2049ms       | 1780.3ms      | 228.3ms              |          |
| #4169 | 6947ms       | 2827.0ms      | 4101.8ms             |          |
| New   | 1049ms       | 838.6ms       | 200.3ms              | 990KB    |


### http://dodona.ugent.be/nl/courses/359/
22 series | 90 exercises | 623 users | 355.789 submissions

| Pr    | Total request time | Time in views | Time in Activerecord | Filesize |
|-------|--------------|---------------|----------------------|----------|
| Old   | 2006ms       | 394.3ms       | 1603.1ms             | 296 KB   |
| #4167 | 1931ms       | 463.9ms       | 1456.2ms             |          |
| #4169 | 1879ms       | 684.6ms       | 1182.8ms             |          |
| New   | 970ms        | 352.4ms       | 607.9ms              | 326.3KB  |

### Notes
The speed up is much less noticeable for the first course. This is due to the first two series of that course being exams in which few students participated. 
In the second course the speedup is much greater then the slowdown, caused by loading the activity names.

In order to limit the increase in file size, I tried to keep the amount of html for each row to a minimum.

## Visual
Old
![image](https://user-images.githubusercontent.com/21177904/202487015-9d5399c6-ad1c-41ae-ad9a-65d42f943406.png)

New
![image](https://user-images.githubusercontent.com/21177904/203263584-c805d3d5-4d9a-4d35-917f-a07ad57f82e8.png)

Closes #3962 .

PS: this pr also containes some changes to the generated database file.
This is probably caused by me having a backup of production on my machine.
I would keep these changes in as they reflect the state of the production database.