@jorg-vr 2022-11-22
@bmesuere 
1. Should be fixed now. Forgot to correctly set the `SKELETON_TABLE_SELECTOR` after testing
2. I switched back to the old height. The height difference was to match the height of the elements in the actual table